fix(stdlib/timer): use self._id for CT dispatch and uint16 for prescaler
All five Timer methods (start, stop, clear, set_compare, overflow) were
using `match self._n:` which reads a numeric SRAM member and generates
runtime comparisons for every timer-number case. Switch to `match self._id:`
(string member = CT constant, same pattern as irq()) so the compiler folds
the dispatch at compile time and emits only the instructions for the selected
timer.

Also fixes `prescaler: uint8` -> `uint16` in Timer.__init__ and all three
timer_init functions: 256 and 1024 overflow uint8, causing TCCR0B/TCCR1B to
be written with 0 (no clock source) instead of the correct prescaler bits.

+++ lib/src/pymcu/hal/_timer/atmega328p.py
@@ -7,7 +7,7 @@ from pymcu.types import uint8, uint16, inline, compile_isr, Callable
 # ---- Timer0 (8-bit, shared with delay_ms / PWM OC0A/OC0B) ----
 
 @inline
-def timer0_init(prescaler: uint8):
+def timer0_init(prescaler: uint16):
     TCCR0A.value = 0x00
     if prescaler == 1:
         TCCR0B.value = 0x01
@@ -57,7 +57,7 @@ def timer0_stop_ctc():
 # OVF vector: 0x000d (word addr); ~0.5 Hz at 16 MHz, prescaler 1024, 16-bit wrap
 
 @inline
-def timer1_init(prescaler: uint8):
+def timer1_init(prescaler: uint16):
     TCCR1A.value = 0x00
     TCCR1B.value = 0x00
     if prescaler == 1:
@@ -111,7 +111,7 @@ def timer1_stop_ctc():
 # OVF vector: 0x0009 (word addr)
 
 @inline
-def timer2_init(prescaler: uint8):
+def timer2_init(prescaler: uint16):
     TCCR2A.value = 0x00
     TCCR2B.value = 0x00
     if prescaler == 1:
